[#71] refactor:보관함 알림 컴포넌트 분리
차후 알림 컴포넌트 내에서도 분리 예정
